refactor: reorder userpage components
