+++ frontend/src/pages/user/CurriculumListRow.tsx
@@ -1,4 +1,4 @@
-import { LowDetailPlanView } from '../client'
+import { LowDetailPlanView } from '../../client'
 
 const CurriculumListRow = ({ handleDelete, curriculum }: { handleDelete: Function, curriculum: LowDetailPlanView }): JSX.Element => {
   function getDateString (date: string): string {
